Hash -> size_t

+++ graph/Graph.cc
@@ -699,7 +699,7 @@ Graph::makeArcDelayPools(ArcIndex arc_count,
     }
 
     // Leave some room for edits.
-    unsigned annot_size = arc_count * 1.2;
+    int annot_size = arc_count * 1.2;
     arc_delay_annotated_.resize(annot_size * ap_count);
   }
 }
@@ -731,9 +731,9 @@ Graph::makeEdgeArcDelays(Edge *edge)
     }
     edge->setArcDelays(arc_index);
     // Make sure there is room for delay_annotated flags.
-    unsigned max_annot_index = (arc_index + arc_count) * ap_count_;
+    size_t max_annot_index = (arc_index + arc_count) * ap_count_;
     if (max_annot_index >= arc_delay_annotated_.size()) {
-      unsigned size = max_annot_index * 1.2;
+      size_t size = max_annot_index * 1.2;
       arc_delay_annotated_.resize(size);
     }
     removeDelayAnnotated(edge);
@@ -817,7 +817,7 @@ Graph::arcDelayAnnotated(Edge *edge,
 			 DcalcAPIndex ap_index) const
 {
   if (arc_delay_annotated_.size()) {
-    unsigned index = (edge->arcDelays() + arc->index()) * ap_count_ + ap_index;
+    size_t index = (edge->arcDelays() + arc->index()) * ap_count_ + ap_index;
     if (index >= arc_delay_annotated_.size())
       internalError("arc_delay_annotated array bounds exceeded");
     return arc_delay_annotated_[index];
@@ -832,7 +832,7 @@ Graph::setArcDelayAnnotated(Edge *edge,
 			    DcalcAPIndex ap_index,
 			    bool annotated)
 {
-  unsigned index = (edge->arcDelays() + arc->index()) * ap_count_ + ap_index;
+  size_t index = (edge->arcDelays() + arc->index()) * ap_count_ + ap_index;
   if (index >= arc_delay_annotated_.size())
     internalError("arc_delay_annotated array bounds exceeded");
   arc_delay_annotated_[index] = annotated;
@@ -843,7 +843,7 @@ Graph::wireDelayAnnotated(Edge *edge,
 			  const TransRiseFall *tr,
 			  DcalcAPIndex ap_index) const
 {
-  unsigned index = (edge->arcDelays() + TimingArcSet::wireArcIndex(tr)) * ap_count_
+  size_t index = (edge->arcDelays() + TimingArcSet::wireArcIndex(tr)) * ap_count_
     + ap_index;
   if (index >= arc_delay_annotated_.size())
     internalError("arc_delay_annotated array bounds exceeded");
@@ -856,7 +856,7 @@ Graph::setWireDelayAnnotated(Edge *edge,
 			     DcalcAPIndex ap_index,
 			     bool annotated)
 {
-  unsigned index = (edge->arcDelays() + TimingArcSet::wireArcIndex(tr)) * ap_count_
+  size_t index = (edge->arcDelays() + TimingArcSet::wireArcIndex(tr)) * ap_count_
     + ap_index;
   if (index >= arc_delay_annotated_.size())
     internalError("arc_delay_annotated array bounds exceeded");
